requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.14
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-ckeditor==6.4.2
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
stripe==3.4.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
